backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
python-dotenv
pydantic-settings
pandas
openai
python-multipart
tenacity
email-validator
